DEP-3909 add PrintUtils class and showTranscriptAndSurvey fn
